tests/requirements.txt
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pyyaml>=6.0.0
rich>=13.7.0
click>=8.1.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
